refactor(splitwise): swap Durable Object for Workers KV
Cloudflare Pages can't host a Durable Object class in-project (the
build step rejects [[migrations]] and requires a script_name pointing
at a separate Worker). KV has none of that ceremony — each group is
one JSON blob under `group:<uuid>`, read-modify-write on every
mutation.

Drops functions/group-do.ts, the [[path]] catch-all, and wrangler.toml
entirely. Restores the per-resource endpoint layout (groups, people,
expenses) backed by KV via new loadGroup/saveGroup helpers in
functions/api/_utils.ts. README now documents the single KV-namespace
setup (create + dashboard binding) plus the eventual-consistency and
shared-namespace caveats.
