--- wrangler.toml
@@ -1,15 +1,0 @@
-name = "taytays-tools"
-compatibility_date = "2025-04-01"
-pages_build_output_dir = "dist"
-
-# Each group lives in its own Durable Object instance, keyed by its secret UUID.
-# Checking this file in is what tells Cloudflare Pages to register the GroupDO
-# class + bind it as `GROUPS` at build time. Without it the class never shows
-# up in the dashboard's Durable Object bindings dropdown.
-[[durable_objects.bindings]]
-name = "GROUPS"
-class_name = "GroupDO"
-
-[[migrations]]
-tag = "v1"
-new_classes = ["GroupDO"]
